requirements.txt
flask~=2.1.2
flask_sqlalchemy
flask_migrate
flask_mail
Werkzeug~=2.1.2
alembic~=1.8.0
SQLAlchemy~=1.4.37
flask_expects_json
python-dotenv
itsdangerous~=2.1.2
gunicorn
psycopg2
flask-swagger
flask-swagger-ui
PyJWT~=2.4.0
